Ticket #3941: Study w/ Invalid Birth Date goes into infinite loop with reconcilliation
Fixed comparison code in StudyComparer. Use DicomAttrubyte.ToString() to retrieve the value in the dicom message for comparison. The same method is used when the database is populated.

Modified RenameFileCommand to backup and restore the source file when rolling back.

Added OpValidationCommand to the validate the result of the operation. It checks if the file has been copied properly to the Reconcile folder and triggers restore of the files.

Added method to simulate errors/exceptions. In DEBUG mode, errors are simulated at the end of RenameFileCommand and MoveReconcileImageCommand.

Modified StudyProcessItemProcessor to log the information about the SOP that is being reconciled, and list the problems that causes the reconciliation.

+++ ImageServer/Common/CommandProcessor/CopyFileCommand.cs
@@ -29,16 +29,20 @@
 
 #endregion
 
+using System;
 using System.IO;
 using ClearCanvas.Common;
 
 namespace ClearCanvas.ImageServer.Common.CommandProcessor
 {
-	public class CopyFileCommand : ServerCommand
+	public class CopyFileCommand : ServerCommand, IDisposable
 	{
-		#region Private Members
+        #region Private Members
 		private readonly string _sourceFile;
 		private readonly string _destinationFile;
+	    private bool _overwrite;
+	    private string _destBackupFile;
+	    private readonly bool _destExists;
 		#endregion
 
 		public CopyFileCommand(string sourceFile, string destinationFile)
@@ -46,25 +50,71 @@ namespace ClearCanvas.ImageServer.Common.CommandProcessor
 		{
 			Platform.CheckForNullReference(sourceFile, "Source filename");
 			Platform.CheckForNullReference(destinationFile, "Destination filename");
+		    Platform.CheckTrue(File.Exists(sourceFile), String.Format("Source file {0} exists", sourceFile));
 
 			_sourceFile = sourceFile;
 			_destinationFile = destinationFile;
+		    _destExists = File.Exists(_destinationFile);
 		}
 
-		protected override void OnExecute()
+	    public bool Overwrite
+	    {
+	        get { return _overwrite; }
+	        set { _overwrite = value; }
+	    }
+
+	    protected override void OnExecute()
 		{
+            if (RequiresRollback)
+                Backup();
+
 			if (File.Exists(_destinationFile))
 			{
 				File.Delete(_destinationFile);
 			}
 
-			File.Copy(_sourceFile, _destinationFile);
+            File.Copy(_sourceFile, _destinationFile, Overwrite);
 		}
 
-		protected override void OnUndo()
+	    private void Backup()
+	    {
+	        // backup the destination
+	        _destBackupFile = _destinationFile + ".bak";
+	        File.Copy(_destinationFile, _destBackupFile, true);
+	    }
+
+	    protected override void OnUndo()
 		{
-			if (File.Exists(_destinationFile))
-				File.Delete(_destinationFile);
+            if (RequiresRollback)
+            {
+                if (!_destExists)
+                {
+                    if (File.Exists(_destinationFile))
+                    {
+                        Platform.Log(LogLevel.Info, "Restoring {0}..", _destinationFile);
+                        File.Delete(_destinationFile);
+                    }
+                }
+                else
+                {
+                    if (File.Exists(_destBackupFile))
+                    {
+                        Platform.Log(LogLevel.Info, "Restoring {0}..", _destinationFile);
+                        File.Copy(_destinationFile, _destBackupFile, true);
+                    }
+                }
+            }
 		}
-	}
+
+        #region IDisposable Members
+
+        public void Dispose()
+        {
+            if (File.Exists(_destBackupFile))
+                File.Delete(_destBackupFile);
+
+        }
+
+        #endregion
+    }
 }
